refactor: relocate server-layer TypedDicts from core/types to server/tools/_types
Move 7 public TypedDicts (KitchenStatusResult, TokenSummaryResult,
TimingSummaryResult, RunSkillResult, RunCmdResult, TestCheckResult,
MergeWorktreeResult) and their 3 private base classes from
core/types/_type_results.py to server/tools/_types.py. ModelTotalEntry
remains in core since it's used by _type_protocols_logging.py.

Update tests/infra/conftest.py to import from new location. Add new
test file tests/server/test_tools_types_module.py for import verification.
