+++ src/autoskillit/server/tools/CLAUDE.md
@@ -7,6 +7,7 @@ MCP `@mcp.tool()` handlers registered on import (14 tool modules).
 | File | Purpose |
 |------|---------|
 | `__init__.py` | Docstring-only — tools register via `@mcp.tool()` on import |
+| `_types.py` | TypedDict definitions for server tool JSON responses (RunSkillResult, RunCmdResult, etc.) |
 | `tools_kitchen.py` | `open_kitchen`, `close_kitchen` (gate lifecycle), `recipe://` MCP resource |
 | `tools_ci.py` | `set_commit_status`, `check_repo_merge_state` |
 | `tools_ci_watch.py` | `wait_for_ci`, `get_ci_status`, `_auto_trigger_ci` |
